requirements.txt
absl-py==0.13.0
astor==0.8.1
backcall==0.2.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cycler==0.11.0
dataclasses==0.6
debugpy==1.7.0
decorator==5.1.1
entrypoints==0.4
Flask==1.1.2
fonttools==4.38.0
gast==0.2.2
google-auth==1.31.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.38.0
h5py==2.10.0
idna==2.10
importlib-metadata==4.5.0
ipykernel==6.16.2
ipython==7.34.0
itsdangerous==2.0.1
jedi==0.19.2
Jinja2==3.0.1
joblib==0.14.1
jupyter_client==7.4.9
jupyter_core==4.12.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.18.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==24.0
pandas==1.1.5
parso==0.8.4
pickle-mixin==1.0.2
pickleshare==0.7.5
Pillow==7.2.0
prompt_toolkit==3.0.48
protobuf==3.11.3
psutil==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.17.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.1
pywin32==308
pyzmq==26.2.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.12.2
six==1.14.0
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
tornado==6.2
traitlets==5.9.0
typing-extensions==3.7.4.3
urllib3==1.26.5
wcwidth==0.2.13
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.4.1
